+++ service/controller/azuremachinepool/handler/nodepool/create.go
@@ -62,7 +62,7 @@ func (r *Resource) EnsureCreated(ctx context.Context, obj interface{}) error {
 		r.Logger.Debugf(ctx, "new state: %s", newState)
 		r.Logger.Debugf(ctx, "setting resource status to %#q", newState)
 		err = r.saveCurrentState(ctx, azureMachinePool, string(newState))
-		if apierrors.IsConflict(err) {
+		if apierrors.IsConflict(microerror.Cause(err)) {
 			r.Logger.Debugf(ctx, "conflict trying to save object in k8s API concurrently")
 			r.Logger.Debugf(ctx, "no state change")
 			return nil