+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Consider node pools out of date if flatcar image is changed.
+
 ## [5.7.0] - 2021-05-13
 
 ### Changed